requirements.txt
streamlit
moviepy==2.0.0.dev10
Pillow
requests
google-cloud-texttospeech
gTTS
